requirements.txt
flask
pandas
scikit-learn
pymongo
flask-bcrypt
python-dotenv
spotipy
itsdangerous
gunicorn
